implemented logging in

+++ watch.sh
@@ -6,7 +6,7 @@ build_project () {
 
     if [[ ! "$?" -eq 0 ]]; then
         echo "failed to build project"
-        exit -1
+        exit 1
     fi
 }
 run_project () {
@@ -24,14 +24,21 @@ inotifywait -m . -r -e "MODIFY" --exclude "(_build/*|.git/*|.*/\.#.*|\./.*\.db|\
     while read -r directory events filename; do
         echo "received event $events on $directory:$filename, restarting"
 
-        build_project
+        if opam exec -- dune build @all; then
 
-        opam exec -- dune build @all
+            while (ps -p "$BG_PID" > /dev/null) && [[ ! -z "$BG_PID" ]]; do
+                echo kill -INT $BG_PID
+                kill -INT $BG_PID
+            done
 
-        while (ps -p "$BG_PID" > /dev/null) && [[ ! -z "$BG_PID" ]]; do
-            echo kill -INT $BG_PID
-            kill -INT $BG_PID
-        done
-        run_project
-        BG_PID="$!"
+            while lsof -i TCP | grep -q 7331; do
+                BAD_PID=$(lsof -i TCP | grep -v 7331 | cut -d ' ' -f 3)
+                echo "Killing $BAD_PID"
+                kill -9 "$BAD_PID"
+            done
+
+            run_project
+
+            BG_PID="$!"
+        fi
     done
